requirements.txt
discord.py>=2.5.0
python-dotenv>=1.0.0
tzdata>=2024.1
